Add scraper company, Categories, and Category files  Description:  This commit adds the following files to the repository:  scraperCompany: This file contains the code for scraping company data. scraperCategories: This file contains the code for scraping all categories data. scraperCategory: This file contains the code for scraping data for a specific category.
Add scraper company, Categories, and Category files

Description:

This commit adds the following files to the repository:

scraperCompany: This file contains the code for scraping company data.
scraperCategories: This file contains the code for scraping all categories data.
scraperCategory: This file contains the code for scraping data for a specific category.